requirements.txt
antlr4-python3-runtime==4.9.3
attrs==25.3.0
Automat==25.4.16
certifi==2025.4.26
charset-normalizer==3.4.2
constantly==23.10.4
construct==2.10.70
contourpy==1.3.2
cycler==0.12.1
einops==0.8.1
filelock==3.18.0
fonttools==4.58.0
fsspec==2025.5.0
hf-xet==1.1.2
huggingface-hub==0.31.4
hydra-core==1.3.2
hyperlink==21.0.0
idna==3.10
imageio==2.37.0
incremental==24.7.2
Jinja2==3.1.6
joblib==1.5.1
kiwisolver==1.4.8
lazy_loader==0.4
MarkupSafe==3.0.2
matplotlib==3.10.3
monai==1.3.0
mpmath==1.3.0
networkx==3.4.2
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
packaging==25.0
pillow==11.2.1
protobuf==6.31.0
pyasn1==0.6.1
pyparsing==3.2.3
pysnmp==7.1.20
python-dateutil==2.9.0.post0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.3
six==1.17.0
sympy==1.14.0
threadpoolctl==3.6.0
tifffile==2025.5.10
tokenizers==0.15.2
tomli==2.2.1
torch==2.2.1
tqdm==4.67.1
transformers==4.39.1
triton==2.2.0
Twisted==24.11.0
typing_extensions==4.13.2
urllib3==2.4.0
zope.interface==7.2
